Add integration source as delivered by Yager.
This is the original source code to the FaceFX / Unreal Engine 4 integration as
delivered by Yager. The source is organized into a UE4 module and there are
several UE4 source code modifications. This code forms the starting point for
the FaceFX / UE4 plugin development. The FaceFX Runtime is not included in this
code -- it is under a separate license and will be distributed separately.