Initial version of the plugin based FaceFX integration for UE4
The previous module based integration was converted into a self contained plugin.
A test map with the example Slade assets are located inside the \UE4\Engine\Engine\Plugins\Runtime\FaceFX\Content folder.
Some parts of the code do not conform to the OC3/EPIC naming conventions yet - that's WIP and will be part of a future CL.