Only reset asset on reimport if all data present
The asset was being reset, clearing out all FaceFX Runtime data in the actor
before it was known if all data is present and a reimport can happen
successfully. If there was missing data, this would result in your current
asset being completely empty and it would either not work or crash. Now the
asset is only reset if all the required data is present during reimport. Only
after verifying this is the asset reset and reimported into.