Change LoadAudioMapData create messages to modify
If the audio map file is not present, fails to load, or has some other error,
this should not be a "create" error that allows the user to roll back the
asset (in this case the actor, destroying it). This commit changes those
messages to modify messages which can't be rolled back.