Fix Rocket build
The build files have been modified so that the GitHub source build and the
"rocket" build use the exact same files (no editing necessary). In addition,
a FilterPlugin.ini file was added so that the rocket build copies over the
readme, license, and config files and deletes the intermediate and source
directories from the plugin package it creates.