Add Mac support
Mac support was added to the build configuration. In addition, several issues
were fixed in the code as a result of compiling with Clang. Finally, the FaceFX
Runtime header facefx.h is only included once now, and it is no longer included
via a #define (which didn't work with UBT's Mac build chain -- it dropped the
quotes).