Fix nonunity compile
Fix FaceFX to compile in NonUnity configuration, and enable bEnforceIWYU in the Build.cs files in order to continue enforcing it.

Unreal expects plugins in the Engine directory to conform to the IWYU standards as of 4.16.

https://docs.unrealengine.com/latest/INT/Programming/UnrealBuildSystem/IWYUReferenceGuide/

Also, compiling with

    <bUseUnityBuild>false</bUseUnityBuild>

in BuildConfiguration.xml helps to verify that all files compile on their own, without depending on being bundled into a unity blob with other cpp files.