Throw a BuildException if the build can't continue
If the FaceFX Runtime directory can't be found, the directory containing
the library that needs to be linked can't be found, or the target
platform is not supported, immediately fail the build by throwing a
BuildException. Previously only a message was printed and lost in the
noise of normal build output with the build continuing resulting in
either a) linker errors or b) a broken build at runtime.