Remove Matinee as part of UE 4.24 support
As of 4.23 Matinee is no longer officially supported and is going to be
removed. The FaceFX Plugin is therefore dropping support for Matinee as
part of its 4.24 support (it has been difficult to keep it building and
working up to this point and this is a logical time to officially drop
it on our end as well). If you need Matinee moving forward, stick with
4.23 or use 4.23's Matinee to Sequencer Conversion Tool and try to move
to Sequencer before updating to 4.24.

fixes: #165
see: #166