Unify FFaceFXEditorModule::PreSaveWorld & OnEndPie
This patch unifies the implementation of
FFaceFXEditorModule::PreSaveWorld() and FFaceFXEditorModule::OnEndPie()
and standardizes on the FFaceFXEditorModule::OnEndPie() approach.

The code in FFaceFXEditorModule::PreSaveWorld() contained a bug in that
the AActor::GetComponents() call used UFaceFXCharacter::StaticClass() as
the first parameter. The first parameter is of type
TSubclassOf<UActorComponent> and UFaceFXCharacter is not such a
subclass.

fixes: #168